:truck: Moved code from progressBar index to utils.tsx
Moved the code to utils so that it can be used by topProgressBar
